User Management refactored

+++ backend/user_management/FriendRequest/models.py
@@ -6,7 +6,14 @@ class FriendRequest(models.Model):
     receiver = models.ForeignKey(Profile, on_delete=models.CASCADE, related_name='received_requests')
     created_at = models.DateTimeField(auto_now_add=True)
 
+    class Meta:
+        unique_together = ['sender', 'receiver']
+        verbose_name = 'Friend Request'
+        verbose_name_plural = 'Friend Requests'
+
     def accept(self):
+        if self.sender.friends.filter(id=self.receiver.id).exists():
+            raise ValueError("Users are already friends.")
         self.sender.friends.add(self.receiver)
         self.receiver.friends.add(self.sender)
         self.delete()
